Replace ParalayoutStudio with new demo app
ParalayoutStudio, the old sample code for experimenting with Paralayout, was primarily focused on functionality around Label, which is being removed from the framework (#19). The project was also outdated (lacked support for iPhone X even) and was not set up with a dependency manager, which made adding snapshot tests difficult.

This replaces ParalayoutStudio with a new demo app that uses CocoaPods for dependency management (which inherently tests our CocoaPods integration) and includes a test suite set up to run snapshot tests. My plan is to use this demo app both as a way to show how Paralayout can be used and as a form of integration tests. It currently only has the skeleton for this new design, but it should be easy to add new sample screens. I've added a basic snapshot test for view spreading to kick off building out the snapshot test suite.