Use CGSize to define fixed spacers
This allows reusing a view distribution in vertical or horizontal distributions (for example preferring a horizontal distribution if the views fit, falling back to vertical otherwise) where the spacer has a different value depending on the axis.

The additional static func allows most callsites to remain unchanged and preserves the simplicity of the most common case.