Make view alignment not sensitive to current frame
The .align() method computes an offset, and then applies that offset to the receiver's frame. This makes the result sensitive to the *current* frame of the receiver, rather than being an absolute computation. In certain circumstances, this can result in layout that "jitters" across a rounding threshold on each layout pass, instead of always behaving the same way. For example:

1. A label starts out at (0,0)
2. untransformedFrameOffset returns (0,4.83333339) to vertically center the label (note: just *beyond* the half-pixel boundary at 3x resolution)
3. The offset is applied and rounded up to the nearest pixel: (0,5)
4. On an second layout pass, untransformedFrameOffset now returns (0,-1.666667) since it is subtracting from 5 instead of adding to 0
5. The offset is applied, gets (0,4.83333321) and now gets rounded *down* to (0,4.6666667)

Signed-off-by: Peter Westen <peter@squareup.com>